refactor: minor restructuring
