+++ packages/server-core/src/adapters/shared/ldap/factory.ts
@@ -5,7 +5,7 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-import type { ILdapClient, ILdapClientFactory } from '../../core';
+import type { ILdapClient, ILdapClientFactory } from '../../../core';
 import { LdapClient } from './module';
 import type { LdapClientOptions } from './types';
