+++ packages/server-core/src/adapters/shared/ldap/index.ts
