+++ packages/server-core/src/adapters/shared/ldap/types.ts
