--- packages/server-core/src/adapters/smtp/index.ts
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) 2024.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-export * from './module';
-export * from './singleton';
-export * from './type';
