+++ packages/server-core/src/app/modules/mail/adapter/index.ts
@@ -5,5 +5,5 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './smtp';
+export * from './smtp/smtp';
 export * from './void';
