+++ packages/server-core/src/app/modules/mail/adapter/smtp/module.ts
@@ -1,12 +1,12 @@
 /*
- * Copyright (c) 2022.
+ * Copyright (c) 2022-2025.
  * Author Peter Placzek (tada5hi)
  * For the full copyright and license information,
  * view the LICENSE file that was distributed with this source code.
  */
 
 import { createTransport } from 'nodemailer';
-import type { SMTPClient, SMTPOptions } from './type';
+import type { SMTPClient, SMTPOptions } from './types';
 
 export function createSMTPClient(options?: SMTPOptions | string) : SMTPClient {
     let transport : SMTPClient;
