+++ packages/server-core/src/app/modules/mail/adapter/smtp/types.ts
